Revert "Merge branch 'huyen.v1' into org-management"
This reverts commit 88f5b9ab937bbd9394f310ff59243ca4eaf72cbe, reversing
changes made to 522010404c92d9ad8b4db849d0fd7d0789818193.

+++ src/main/java/fptu/fcharity/entity/Wallet.java
@@ -6,7 +6,6 @@ import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.Nationalized;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -21,6 +20,6 @@ public class Wallet {
 
     @Nationalized
     @Column(name = "balance")
-    private BigDecimal balance;
+    private int balance;
 
 }
