+++ src/main/java/fptu/fcharity/service/WalletService.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Service
@@ -16,7 +15,7 @@ public class WalletService {
     @Transactional
     public Wallet save() {
         Wallet wallet = new Wallet();
-        wallet.setBalance(BigDecimal.valueOf(0.00));
+        wallet.setBalance(0);
         return walletRepository.save(wallet);
     }
 }
