+++ src/main/java/fptu/fcharity/service/manage/organization/OrganizationServiceImpl.java
@@ -18,8 +18,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.io.IOException;
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
@@ -94,7 +92,7 @@ public class OrganizationServiceImpl implements OrganizationService {
         organization.setCeo(ceo);
 
         Wallet wallet = new Wallet();
-        wallet.setBalance(BigDecimal.valueOf(0));
+        wallet.setBalance(0);
         Wallet savedWallet =  walletRepository.save(wallet);
         organization.setWalletAddress(savedWallet);
         organization.setOrganizationStatus(OrganizationStatus.PENDING);
